# actinia_gdi/static/swagger.json
{"consumes": ["application/json"], "schemes": ["https", "http"], "swagger": "2.0", "definitions": {"SimpleResponseModel": {"description": "Simple response schema to inform about status.", "type": "object", "required": ["status", "message"], "properties": {"status": {"description": "The status code of the request.", "type": "int"}, "message": {"description": "A short message to describes the status", "type": "string"}}, "example": {"status": 200, "message": "success"}}, "GeodataResponseModel": {"description": "Model for object for geodata", "type": "object", "required": ["uuid", "bbox"], "properties": {"bbox": {"maxItems": 4, "minItems": 4, "type": "array", "description": "The bounding box of the result.", "items": {"type": "number"}}, "crs": {"description": "The coordinate reference system of the result.", "type": "string"}, "uuid": {"description": "The Geonetwork uuid.", "type": "string"}, "table": {"description": "The database connection string of the source of the result.", "type": "string"}}, "example": {"bbox": [51.1, -34.6, -17.3, 38.2], "crs": "urn:ogc:def:crs:::WGS 1984", "uuid": "da165110-88fd-11da-a88f-000d939bc5d8", "table": "http://www.fao.org/ag/AGL/aglw/aquastat/watresafrica/index.stm"}}}, "info": {"description": "\n                   ", "termsOfService": "", "version": "0.1.0", "title": "actinia-gdi"}, "paths": {"/files": {"post": {"description": "File can be uploaded, best used with https://bmvimetadaten.mundialis.de.", "summary": "Upload file.", "tags": ["file"], "responses": {"200": {"description": "Success or failure of connection", "schema": {"$ref": "#/definitions/SimpleResponseModel"}}}}}, "/metadata/geodata/tags/{tags}": {"get": {"description": "The request will ask Geonetwork which metadata records are available for a certain tag or more tags separated by comma and returns a parsed record build from model. At the moment only the first record is returned.", "summary": "Get geodata object from requests to Geonetwork opensource by one or many tags.", "parameters": [{"description": "One or more Geonetwork tags, comma separated", "name": "tags", "required": true, "type": "string", "in": "path"}], "tags": ["geonetwork"], "responses": {"200": {"description": "Modelled Search Results from Geonetwork", "schema": {"$ref": "#/definitions/GeodataResponseModel"}}}}}, "/metadata/geodata/uuids/{uuid}": {"get": {"description": "The request will ask Geonetwork which metadata records are available for a certain uuid and returns a parsed record build from model. At the moment only the first record is returned.", "summary": "Get geodata object from requests to Geonetwork opensource by uuid.", "parameters": [{"description": "A Geonetwork uuid from a record", "name": "uuid", "required": true, "type": "string", "in": "path"}], "tags": ["geonetwork"], "responses": {"200": {"description": "Modelled Search Results from Geonetwork", "schema": {"$ref": "#/definitions/GeodataResponseModel"}}}}}, "/metadata/raw/categories/{category}": {"get": {"description": "The request will ask Geonetwork which metadata records are available for a certain category and returns the JSON response with these records. Requirement: a virtual CSW is defined in Geonetwork", "summary": "Requests a category from Geonetwork opensource.", "parameters": [{"description": "A Geonetwork category", "name": "category", "required": true, "type": "string", "in": "path"}], "tags": ["geonetwork"], "responses": {"200": {"description": "The Search Results from Geonetwork", "schema": {"example": {"csw:GetRecordsResponse": {"@xmlns:xsi": "http://www.w3.org/2001/XMLSchema-instance", "csw:SearchStatus": {"@timestamp": "2018-03-28T13:46:43"}, "csw:SearchResults": {"@nextRecord": "0", "@elementSet": "full", "@numberOfRecordsMatched": "2", "@numberOfRecordsReturned": "2", "csw:Record": [{"dc:URI": {"@description": "Web page listing Data Service Providers (DSP)s", "#text": "http://www.land.vic.gov.au/vicmapdsp"}, "@xmlns:ows": "http://www.opengis.net/ows", "dc:type": "dataset", "dc:description": "This dataset is the definitive set of locality boundaries for the state of Victoria            as defined by Local Government and registered by the Registrar of Geographic Names.            The boundaries are aligned to Vicmap Property. This dataset is part of the Vicmap Admin dataset series.", "ows:BoundingBox": {"@crs": "urn:ogc:def:crs:::4283", "ows:UpperCorner": "141 -34", "ows:LowerCorner": "150 -39"}, "dc:rights": ["intellectualPropertyRights", "otherRestrictions"], "@xmlns:dc": "http://purl.org/dc/elements/1.1/", "@xmlns:geonet": "http://www.fao.org/geonetwork", "dct:abstract": "This dataset is the definitive set of locality boundaries for the state of Victoria            as defined by Local Government and registered by the Registrar of Geographic Names.            The boundaries are aligned to Vicmap Property. This dataset is part of the Vicmap Admin dataset series.", "dc:source": "Local Government defines suburb, town and rural district boundaries after seeking appropriate consultation with all interested parties.              Proposed boundaries/names are submitted to the Registrar of Geographic Names for approval and formal gazettal.              As boundary and name information changes are processed through the formal definition and gazettal process at the Registrar of Geographic Names, the amendments will be flagged              and the dataset will be updated by the maintainer with all new boundaries aligned to Vicmap Property.", "dc:format": ["Most popular formats including ESRI shape, MapInfo Tab and Oracle Spatial", "Most popular formats including ESRI shape, MapInfo Tab and Oracle Spatial"], "dc:title": "Localities in Victoria (VMADMIN.LOCALITY_POLYGON) - Comprehensive Elements", "dc:date": "2018-02-02T09:47:25", "dc:identifier": "d619aac5-6fc7-44f2-9552-7f8bb4d826e0", "dc:language": "eng", "@xmlns:dct": "http://purl.org/dc/terms/", "dc:subject": ["BOUNDARIES-Administrative", "LAND-Ownership", "boundaries"]}, {"dc:URI": [{"@protocol": "WWW:DOWNLOAD-1.0-http--download", "@description": "Physiography of North and Central Eurasia Landform (Gif Format)", "#text": "http://localhost:8080/geonetwork/srv/api/records/78f93047-74f8-4419-ac3d-fc62e4b0477b/attachments/phy.zip", "@name": "phy.zip"}, {"@protocol": "OGC:WMS", "@description": "Physiography of North and Central Eurasia Landform", "#text": "http://data.fao.org/maps/wms", "@name": "GEONETWORK:phy_landf_7386"}, {"@protocol": "OGC:WMS", "@description": "Physiography of North and Central Eurasia Slope", "#text": "http://data.fao.org/maps/wms", "@name": "GEONETWORK:phy_slope_7386"}, {"@protocol": "OGC:WMS", "@description": "Physiography of North and Central Eurasia Hypsography", "#text": "http://data.fao.org/maps/wms", "@name": "GEONETWORK:phy_hypso_7386"}, {"@protocol": "image/gif", "#text": "http://localhost:8080/geonetwork/srv/api/records/78f93047-74f8-4419-ac3d-fc62e4b0477b/attachments/phy_s.gif", "@name": "Physiographic Map (small preview)"}, {"@protocol": "image/gif", "#text": "http://localhost:8080/geonetwork/srv/api/records/78f93047-74f8-4419-ac3d-fc62e4b0477b/attachments/phy.gif", "@name": "Physiographic Map"}], "@xmlns:ows": "http://www.opengis.net/ows", "@xmlns:dct": "http://purl.org/dc/terms/", "dc:description": "Physiographic maps for the CIS and Baltic States (CIS_BS), Mongolia, China and Taiwan Province of China. Between the three regions (China, Mongolia, and CIS_BS countries) DCW boundaries were introduced. There are no DCW boundaries between Russian Federation and the rest of the new countries of the CIS_BS. The original physiographic map of China includes the Chinese border between India and China, which extends beyond the Indian border line, and the South China Sea islands (no physiographic information is present for islands in the South China Sea). The use of these country boundaries does not imply the expression of any opinion whatsoever on the part of FAO concerning the legal or constitutional states of any country, territory, or sea area, or concerning delimitation of frontiers. The Maps visualize the items LANDF, HYPSO, SLOPE that correspond to Landform, Hypsometry and Slope.", "ows:BoundingBox": {"@crs": "urn:ogc:def:crs:::Lambert Azimuthal Projection", "ows:UpperCorner": "37 83", "ows:LowerCorner": "156 -3"}, "dc:rights": ["copyright", "copyright"], "@xmlns:dc": "http://purl.org/dc/elements/1.1/", "@xmlns:geonet": "http://www.fao.org/geonetwork", "dct:abstract": "Physiographic maps for the CIS and Baltic States (CIS_BS), Mongolia, China and Taiwan Province of China. Between the three regions (China, Mongolia, and CIS_BS countries) DCW boundaries were introduced. There are no DCW boundaries between Russian Federation and the rest of the new countries of the CIS_BS. The original physiographic map of China includes the Chinese border between India and China, which extends beyond the Indian border line, and the South China Sea islands (no physiographic information is present for islands in the South China Sea). The use of these country boundaries does not imply the expression of any opinion whatsoever on the part of FAO concerning the legal or constitutional states of any country, territory, or sea area, or concerning delimitation of frontiers. The Maps visualize the items LANDF, HYPSO, SLOPE that correspond to Landform, Hypsometry and Slope.", "dc:source": null, "dc:title": "Physiographic Map of North and Central Eurasia (Sample record, please remove!)", "dc:date": "2018-02-02T09:47:45", "dc:identifier": "78f93047-74f8-4419-ac3d-fc62e4b0477b", "dc:language": null, "dc:subject": ["physiography, soil", "Eurasia", "geoscientificInformation"]}]}, "@xmlns:csw": "http://www.opengis.net/cat/csw/2.0.2", "@xsi:schemaLocation": "http://www.opengis.net/cat/csw/2.0.2 http://schemas.opengis.net/csw/2.0.2/CSW-discovery.xsd"}}}}}}}, "/metadata/raw/tags/{tags}": {"get": {"description": "The request will ask Geonetwork which metadata records are available for a certain tag or more tags separated by comma and returns the JSON response with these records.", "summary": "Requests one or many tags from Geonetwork opensource.", "parameters": [{"description": "One or more Geonetwork tags, comma separated", "name": "tags", "required": true, "type": "string", "in": "path"}], "tags": ["geonetwork"], "responses": {"200": {"description": "The Search Results from Geonetwork", "schema": {"example": {"csw:GetRecordsResponse": {"@xmlns:xsi": "http://www.w3.org/2001/XMLSchema-instance", "csw:SearchStatus": {"@timestamp": "2018-03-28T13:42:41"}, "csw:SearchResults": {"@nextRecord": "0", "@elementSet": "full", "@numberOfRecordsMatched": "2", "@numberOfRecordsReturned": "2", "csw:Record": [{"dc:URI": [{"@protocol": "WWW:LINK-1.0-http--link", "@description": "Online link to the  'Water Resources and Irrigation in Africa'- website", "#text": "http://www.fao.org/ag/AGL/aglw/aquastat/watresafrica/index.stm", "@name": ""}, {"@protocol": "WWW:DOWNLOAD-1.0-http--download", "@description": "Hydrological basins in Africa (Shapefile Format)", "#text": "http://localhost:8080/geonetwork/srv/api/records/da165110-88fd-11da-a88f-000d939bc5d8/attachments/basins.zip", "@name": "basins.zip"}, {"@protocol": "OGC:WMS", "@description": "Hydrological basins in Africa", "#text": "http://data.fao.org/maps/wms", "@name": "GEONETWORK:basins_296"}, {"@protocol": "image/gif", "#text": "http://localhost:8080/geonetwork/srv/api/records/da165110-88fd-11da-a88f-000d939bc5d8/attachments/thumbnail_s.gif", "@name": "Hydrological Basins in Africa (small preview)"}, {"@protocol": "image/gif", "#text": "http://localhost:8080/geonetwork/srv/api/records/da165110-88fd-11da-a88f-000d939bc5d8/attachments/thumbnail.gif", "@name": "Hydrological Basins in Africa"}], "@xmlns:ows": "http://www.opengis.net/ows", "@xmlns:dct": "http://purl.org/dc/terms/", "dc:description": "Major hydrological basins and their sub-basins. This dataset divides the African continent according to its hydrological characteristics.\nThe dataset consists of the following information:- numerical code and name of the major basin (MAJ_BAS and MAJ_NAME); - area of the major basin in square km (MAJ_AREA); - numerical code and name of the sub-basin (SUB_BAS and SUB_NAME); - area of the sub-basin in square km (SUB_AREA); - numerical code of the sub-basin towards which the sub-basin flows (TO_SUBBAS) (the codes -888 and -999 have been assigned respectively to internal sub-basins and to sub-basins draining into the sea)", "ows:BoundingBox": {"@crs": "urn:ogc:def:crs:::WGS 1984", "ows:UpperCorner": "-17.3 38.2", "ows:LowerCorner": "51.1 -34.6"}, "@xmlns:dc": "http://purl.org/dc/elements/1.1/", "@xmlns:geonet": "http://www.fao.org/geonetwork", "dct:abstract": "Major hydrological basins and their sub-basins. This dataset divides the African continent according to its hydrological characteristics.\nThe dataset consists of the following information:- numerical code and name of the major basin (MAJ_BAS and MAJ_NAME); - area of the major basin in square km (MAJ_AREA); - numerical code and name of the sub-basin (SUB_BAS and SUB_NAME); - area of the sub-basin in square km (SUB_AREA); - numerical code of the sub-basin towards which the sub-basin flows (TO_SUBBAS) (the codes -888 and -999 have been assigned respectively to internal sub-basins and to sub-basins draining into the sea)", "dc:source": "The linework of the map is obtained by delineating drainage basin boundaries from an hydrologically corrected digital elevation model with a resolution of 1 * 1 km.", "dc:format": ["ShapeFile", "ShapeFile"], "dc:title": "Hydrological Basins in Africa (Sample record, please remove!)", "dc:date": "2018-02-02T09:46:51", "dc:identifier": "da165110-88fd-11da-a88f-000d939bc5d8", "dc:language": "eng", "dc:subject": ["watersheds", "river basins", "water resources", "hydrology", "AQUASTAT", "AWRD", "Africa", "inlandWaters"]}, {"dc:URI": {"@description": "Web page listing Data Service Providers (DSP)s", "#text": "http://www.land.vic.gov.au/vicmapdsp"}, "@xmlns:ows": "http://www.opengis.net/ows", "dc:type": "dataset", "dc:description": "This dataset is the definitive set of locality boundaries for the state of Victoria            as defined by Local Government and registered by the Registrar of Geographic Names.            The boundaries are aligned to Vicmap Property. This dataset is part of the Vicmap Admin dataset series.", "ows:BoundingBox": {"@crs": "urn:ogc:def:crs:::4283", "ows:UpperCorner": "141 -34", "ows:LowerCorner": "150 -39"}, "dc:rights": ["intellectualPropertyRights", "otherRestrictions"], "@xmlns:dc": "http://purl.org/dc/elements/1.1/", "@xmlns:geonet": "http://www.fao.org/geonetwork", "dct:abstract": "This dataset is the definitive set of locality boundaries for the state of Victoria            as defined by Local Government and registered by the Registrar of Geographic Names.            The boundaries are aligned to Vicmap Property. This dataset is part of the Vicmap Admin dataset series.", "dc:source": "Local Government defines suburb, town and rural district boundaries after seeking appropriate consultation with all interested parties.              Proposed boundaries/names are submitted to the Registrar of Geographic Names for approval and formal gazettal.              As boundary and name information changes are processed through the formal definition and gazettal process at the Registrar of Geographic Names, the amendments will be flagged              and the dataset will be updated by the maintainer with all new boundaries aligned to Vicmap Property.", "dc:format": ["Most popular formats including ESRI shape, MapInfo Tab and Oracle Spatial", "Most popular formats including ESRI shape, MapInfo Tab and Oracle Spatial"], "dc:title": "Localities in Victoria (VMADMIN.LOCALITY_POLYGON) - Comprehensive Elements", "dc:date": "2018-02-02T09:47:25", "dc:identifier": "d619aac5-6fc7-44f2-9552-7f8bb4d826e0", "dc:language": "eng", "@xmlns:dct": "http://purl.org/dc/terms/", "dc:subject": ["BOUNDARIES-Administrative", "LAND-Ownership", "boundaries"]}]}, "@xmlns:csw": "http://www.opengis.net/cat/csw/2.0.2", "@xsi:schemaLocation": "http://www.opengis.net/cat/csw/2.0.2 http://schemas.opengis.net/csw/2.0.2/CSW-discovery.xsd"}}}}}}}, "/metadata/raw/uuids/{uuid}": {"get": {"description": "The request will ask Geonetwork which metadata records are available for a certain uuid and returns the JSON response with this record.", "summary": "Requests an uuid from Geonetwork opensource.", "parameters": [{"description": "A Geonetwork uuid from a record", "name": "uuid", "required": true, "type": "string", "in": "path"}], "tags": ["geonetwork"], "responses": {"200": {"description": "The Search Results from Geonetwork", "schema": {"example": {"csw:GetRecordByIdResponse": {"@xmlns:csw": "http://www.opengis.net/cat/csw/2.0.2", "csw:Record": {"dc:URI": [{"@protocol": "WWW:LINK-1.0-http--link", "@description": "Online link to the  'Water Resources and Irrigation in Africa'- website", "#text": "http://www.fao.org/ag/AGL/aglw/aquastat/watresafrica/index.stm", "@name": ""}, {"@protocol": "WWW:DOWNLOAD-1.0-http--download", "@description": "Hydrological basins in Africa (Shapefile Format)", "#text": "http://localhost:8080/geonetwork/srv/api/records/da165110-88fd-11da-a88f-000d939bc5d8/attachments/basins.zip", "@name": "basins.zip"}, {"@protocol": "OGC:WMS", "@description": "Hydrological basins in Africa", "#text": "http://data.fao.org/maps/wms", "@name": "GEONETWORK:basins_296"}, {"@protocol": "image/gif", "#text": "http://localhost:8080/geonetwork/srv/api/records/da165110-88fd-11da-a88f-000d939bc5d8/attachments/thumbnail_s.gif", "@name": "Hydrological Basins in Africa (small preview)"}, {"@protocol": "image/gif", "#text": "http://localhost:8080/geonetwork/srv/api/records/da165110-88fd-11da-a88f-000d939bc5d8/attachments/thumbnail.gif", "@name": "Hydrological Basins in Africa"}], "@xmlns:ows": "http://www.opengis.net/ows", "@xmlns:dct": "http://purl.org/dc/terms/", "dc:description": "Major hydrological basins and their sub-basins. This dataset divides the African continent according to its hydrological characteristics.\nThe dataset consists of the following information:- numerical code and name of the major basin (MAJ_BAS and MAJ_NAME); - area of the major basin in square km (MAJ_AREA); - numerical code and name of the sub-basin (SUB_BAS and SUB_NAME); - area of the sub-basin in square km (SUB_AREA); - numerical code of the sub-basin towards which the sub-basin flows (TO_SUBBAS) (the codes -888 and -999 have been assigned respectively to internal sub-basins and to sub-basins draining into the sea)", "ows:BoundingBox": {"@crs": "urn:ogc:def:crs:::WGS 1984", "ows:UpperCorner": "-17.3 38.2", "ows:LowerCorner": "51.1 -34.6"}, "@xmlns:dc": "http://purl.org/dc/elements/1.1/", "@xmlns:geonet": "http://www.fao.org/geonetwork", "dct:abstract": "Major hydrological basins and their sub-basins. This dataset divides the African continent according to its hydrological characteristics.\nThe dataset consists of the following information:- numerical code and name of the major basin (MAJ_BAS and MAJ_NAME); - area of the major basin in square km (MAJ_AREA); - numerical code and name of the sub-basin (SUB_BAS and SUB_NAME); - area of the sub-basin in square km (SUB_AREA); - numerical code of the sub-basin towards which the sub-basin flows (TO_SUBBAS) (the codes -888 and -999 have been assigned respectively to internal sub-basins and to sub-basins draining into the sea)", "dc:source": "The linework of the map is obtained by delineating drainage basin boundaries from an hydrologically corrected digital elevation model with a resolution of 1 * 1 km.", "dc:format": ["ShapeFile", "ShapeFile"], "dc:title": "Hydrological Basins in Africa (Sample record, please remove!)", "dc:date": "2018-02-02T09:46:51", "dc:identifier": "da165110-88fd-11da-a88f-000d939bc5d8", "dc:language": "eng", "dc:subject": ["watersheds", "river basins", "water resources", "hydrology", "AQUASTAT", "AWRD", "Africa", "inlandWaters"]}}}}}}}}, "/metadata/test/connection": {"get": {"description": "The request will ask the backend if it can successfully connect to Geonetwork.", "summary": "Tests for active connection to Geonetwork opensource.", "tags": ["geonetwork"], "responses": {"200": {"description": "Success or failure of connection", "schema": {"$ref": "#/definitions/SimpleResponseModel"}}}}, "post": {"description": "The request will ask the backend if it can successfully connect to Geonetwork.", "summary": "Tests for active connection to Geonetwork opensource.", "tags": ["geonetwork"], "responses": {"200": {"description": "Success or failure of connection", "schema": {"$ref": "#/definitions/SimpleResponseModel"}}}}}}}